chore: Add flatten option to makeI18nMessagesMap function
This commit adds a new optional parameter `flatten` to the `makeI18nMessagesMap` function in order to support flattening the output object. The `flatten` parameter is set to `false` by default. The changes include modifying the function signature and updating the implementation to conditionally flatten the output object based on the `flatten` parameter value.

```